Update DOCKER_README.md
Capitalization. RAC is always Oracle RAC. GI is not an authorized abbreviation, but I don't know fully where it is used, so I'll leave it for clarity. Example is what you mean (exemplary, hypothetical) not sample (a part of a whole, such as a blood sample), but the error is so common, I don't see the point of correcting it.